Create README.md
Logically Exploring Future of AI development. What is likely to be a Superintelligence? What goals it may have? Will it value biological life? What can even be considered true Superintelligence? It is logical analysis of well known facts and it does lead to interesting conclusions, that if AI reasons well it may indeed be itself interested in protecting biological life for at least one reason: own assurance of existence. Some big challenge is also identified, that is- to ensure it reasons well at all times.